ajustes modulos.
