+++ .idea/misc.xml
@@ -11,6 +11,9 @@
         <entry key="../../../../../../layout/compose-model-1655835438388.xml" value="0.1097972972972973" />
         <entry key="../../../../../../layout/compose-model-1655835589481.xml" value="0.1097972972972973" />
         <entry key="../../../../../../layout/compose-model-1655836212604.xml" value="0.16131756756756757" />
+        <entry key="../../../../../../layout/compose-model-1655837044767.xml" value="0.19609856262833675" />
+        <entry key="../../../../../../layout/compose-model-1655837131110.xml" value="0.16131756756756757" />
+        <entry key="../../../../../../layout/compose-model-1655837603264.xml" value="0.1" />
         <entry key="../../../../.gradle/caches/transforms-3/f985ec0641a7bf7312966e8fb0188a4d/transformed/core-1.7.0/res/drawable/notification_bg_low.xml" value="0.1015" />
       </map>
     </option>
