+++ app/build.gradle
@@ -49,6 +49,9 @@ android {
 
 dependencies {
 
+    implementation project(path: ':data')
+    implementation project(path: ':domain')
+
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.material:material:$compose_version"
@@ -88,9 +91,10 @@ dependencies {
     implementation "androidx.room:room-ktx:$room_version"
     implementation "androidx.room:room-paging:$room_version"
 
-    implementation "androidx.compose.material:material-icons-extended:$compose_version"
-
-
     // Paging 3.0
     implementation 'androidx.paging:paging-compose:1.0.0-alpha14'
+
+//    Icons
+    implementation "androidx.compose.material:material-icons-extended:$compose_version"
+
 }
