+++ app/src/main/java/sistemas/jd/gomes/aeroinfo/MainActivity.kt
@@ -3,13 +3,7 @@ package sistemas.jd.gomes.aeroinfo
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
+import sistemas.jd.gomes.aeroinfo.presentation.list.ListAirport
 import sistemas.jd.gomes.aeroinfo.ui.theme.AeroinfoTheme
 
 class MainActivity : ComponentActivity() {
