+++ app/src/main/java/sistemas/jd/gomes/aeroinfo/presentation/detail/info/NavigationGraph.kt
@@ -1,9 +1,13 @@
-package sistemas.jd.gomes.aeroinfo
+package sistemas.jd.gomes.aeroinfo.presentation.detail.info
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import sistemas.jd.gomes.aeroinfo.presentation.detail.info.charts.InfoCharts
+import sistemas.jd.gomes.aeroinfo.presentation.detail.info.general.InfoGeneral
+import sistemas.jd.gomes.aeroinfo.presentation.detail.info.rotaer.InfoRotaer
+import sistemas.jd.gomes.aeroinfo.presentation.detail.info.weather.InfoWeather
 
 @Composable
 fun NavigationGraph(navController: NavHostController) {
