+++ app/src/main/java/sistemas/jd/gomes/aeroinfo/presentation/detail/info/rotaer/InfoRotaer.kt
@@ -1,4 +1,4 @@
-package sistemas.jd.gomes.aeroinfo
+package sistemas.jd.gomes.aeroinfo.presentation.detail.info.rotaer
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -13,6 +13,7 @@ import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
+import sistemas.jd.gomes.aeroinfo.R
 
 @Composable
 fun InfoRotaer() {
