+++ app/src/main/java/sistemas/jd/gomes/aeroinfo/presentation/detail/info/weather/InfoWeather.kt
@@ -1,0 +1,51 @@
+package sistemas.jd.gomes.aeroinfo.presentation.detail.info.weather
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Divider
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import sistemas.jd.gomes.aeroinfo.R
+import sistemas.jd.gomes.aeroinfo.presentation.list.ListAirportContent
+import sistemas.jd.gomes.aeroinfo.ui.theme.BlueDark
+import sistemas.jd.gomes.aeroinfo.ui.theme.GrayPrimary
+
+@Composable
+fun InfoWeather() {
+    Scaffold(
+        backgroundColor = MaterialTheme.colors.GrayPrimary,
+        contentColor = Color.White,
+        topBar = {
+
+        },
+        content = {
+            Divider(
+                modifier = Modifier
+                    .height(200.dp)
+                    .background(MaterialTheme.colors.BlueDark)
+                    .clip(RoundedCornerShape(20.dp))
+            )
+        }
+    )
+}
+
+@Composable
+@Preview
+private fun ShowScreen() {
+    InfoWeather()
+}
